datasource: configurable default database, less restrictive ad-hoc filters
The datasource now allows default database to be configurable, this is
used in two places:

* In query editor, the database field is prefilled with the default database
* In ad-hoc filters, the database is stripped from column names to make them
  less verbose (e.g. if default database is 'default', it will only show 'table.column' pairs)

The datasource also allows non-fully-qualified column names in the filter.
In such case the database and table will be filled for each target independently,
e.g. if there are two graphs in the dashboard with 'default.table1' and 'default.table2',
and ad-hoc filter is 'column1 = 5', then it will apply in both cases. If the ad-hoc filter
is 'table2.column1 = 5' then it will only apply in the second graph.